pyproject.toml
[project]
name = "SprintAgent"
version = "0.1.0"
description = "Autonomous, fault-tolerant hierarchical agentic system that transforms high-level project briefs into execution-ready roadmaps through a pipeline: **project** → **module** → **task** → **visualization**." 
readme = "README.md"
requires-python = ">=3.11, <3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "colorama>=0.4.6",
    "open-clip-torch>=2.20.0",
    "openai>=2.8.0",
    "openai-agents>=0.6.4",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
